+++ ElvUI_MerathilisUI/Core/Changelog.lua
@@ -13,6 +13,10 @@ local DISABLED_FONT_COLOR = DISABLED_FONT_COLOR
 local ChangeLogData = {
 	"Changes:",
 		"• Fix WeakAuras skinning. Credits: Windtools <3",
+		"• Fix Auctionator Skin",
+		"• Fix MicroBar right click on Guild Icon",
+		"• Added a new Font Outline Option: Shadow (like ElvUI)",
+		"• Added Skin for BagSync",
 
 		-- "• ''",
 
@@ -27,19 +31,6 @@ local ChangeLogData = {
 
 		-- "• ''",
 }
-
-local URL_PATTERNS = {
-	"^(%a[%w+.-]+://%S+)",
-	"%f[%S](%a[%w+.-]+://%S+)",
-	"^(www%.[-%w_%%]+%.(%a%a+))",
-	"%f[%S](www%.[-%w_%%]+%.(%a%a+))",
-	"(%S+@[%w_.-%%]+%.(%a%a+))",
-}
-
-local function formatURL(url)
-	url = "|cff".."149bfd".."|Hurl:"..url.."|h["..url.."]|h|r ";
-	return url
-end
 
 local function ModifiedString(string)
 	local count = find(string, ":")
@@ -59,13 +50,6 @@ local function ModifiedString(string)
 
 	for pattern in gmatch(string, "('.*')") do
 		newString = newString:gsub(pattern, "|cFFFF8800" .. pattern:gsub("'", "") .. "|r")
-	end
-
-	-- find urls
-	for _, v in pairs(URL_PATTERNS) do
-		if find(string, v) then
-			newString = gsub(string, v, formatURL("%1"))
-		end
 	end
 
 	return newString