+++ CHANGELOG.md
@@ -1,3 +1,7 @@
 ### Changes ###
 
   * Fix WeakAuras skinning. Credits: Windtools <3
+  * Fix Auctionator Skin
+  * Fix MicroBar right click on Guild Icon
+  * Added a new Font Outline Option: Shadow (like default ElvUI)
+  * Added Skin for BagSync